pyproject.toml
[build-system]
requires = ["setuptools >= 61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "notabot"
description = "Discord bot for Vectozavr server"
readme = "README.md"
version = "v1.6"
dependencies = [
    "discord.py",
    "asyncj"
]
authors = [{name = "Nakidai", email = "plaza521@inbox.ru"}]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
]

[project.scripts]
notabot = "notabot:main"
